Reformat index.html and main.css
I don't really know what the standard naming conventions for classes and IDs are, but I figure the best way for me to do it, at least for now, is to create class names that essentially work their way down the DOM to the location of the element. This way, I can list all of the general formatting rules for element types in the main.css file, and then all of the classes after that, in order of "depth", so that I can start from the top of the DOM, working my way down to the bottom, and overriding things when nessicary. I don't know if this is standard, but It makes sense to me as a way to keep things simple and consistent, so that's what I've done in this reformat of index.html and styles/main.css.