Fix responsive navigation branch
Previously, the text was unaligned on the mobile version, but it's now fixed. The other problem was that the navigation bar would expand vertically to cover the top of the page's content. To fix this, I added a larger margin into the css media query for mobile devices, which pushes the rest of the page's content further down, to avoid elements overlapping.